requirements.txt
# Core API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-settings>=2.2.1
httpx>=0.27.0
aiofiles>=24.1.0

# LLM & RAG Orchestration
langchain-core>=0.2.33
langchain-text-splitters>=0.2.0   # Used in chunker.py (MarkdownHeader + Recursive)
langgraph>=0.2.9

# Document Parsers
llama-parse>=0.5.2
llama-index-core>=0.11.1
pymupdf>=1.24.9

# Vector DB & Storage
pinecone[grpc]>=5.0.0
supabase>=2.8.1
motor>=3.5.1      # Async MongoDB client

# Auth & Security
Authlib>=1.3.1
PyJWT>=2.9.0
itsdangerous>=2.2.0

# Performance & Observability
upstash-redis>=1.1.0
langfuse>=2.0.0

# Utilities
pybreaker>=1.3.0
tiktoken>=0.7.0     # Transitive dep: langchain-core + llama-index-core token counting

